+++ backend/src/main/java/com/allog/dallog/presentation/MemberController.java
@@ -1,10 +1,10 @@
-package com.allog.dallog.domain.member.presentation;
+package com.allog.dallog.presentation;
 
-import com.allog.dallog.domain.auth.dto.LoginMember;
-import com.allog.dallog.domain.auth.presentation.AuthenticationPrincipal;
 import com.allog.dallog.domain.member.application.MemberService;
+import com.allog.dallog.domain.auth.dto.LoginMember;
 import com.allog.dallog.domain.member.dto.MemberResponse;
 import com.allog.dallog.domain.member.dto.MemberUpdateRequest;
+import com.allog.dallog.presentation.auth.AuthenticationPrincipal;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
