+++ backend/src/main/java/com/allog/dallog/presentation/ScheduleController.java
@@ -1,13 +1,13 @@
-package com.allog.dallog.domain.schedule.presentation;
+package com.allog.dallog.presentation;
 
-import com.allog.dallog.domain.auth.dto.LoginMember;
-import com.allog.dallog.domain.auth.presentation.AuthenticationPrincipal;
 import com.allog.dallog.domain.schedule.application.ScheduleService;
+import com.allog.dallog.domain.auth.dto.LoginMember;
 import com.allog.dallog.domain.schedule.dto.request.DateRangeRequest;
 import com.allog.dallog.domain.schedule.dto.request.ScheduleCreateRequest;
 import com.allog.dallog.domain.schedule.dto.request.ScheduleUpdateRequest;
 import com.allog.dallog.domain.schedule.dto.response.MemberScheduleResponses;
 import com.allog.dallog.domain.schedule.dto.response.ScheduleResponse;
+import com.allog.dallog.presentation.auth.AuthenticationPrincipal;
 import java.net.URI;
 import javax.validation.Valid;
 import org.springframework.http.ResponseEntity;
