+++ backend/src/main/java/com/allog/dallog/presentation/SubscriptionController.java
@@ -1,11 +1,11 @@
-package com.allog.dallog.domain.subscription.presentation;
+package com.allog.dallog.presentation;
 
-import com.allog.dallog.domain.auth.dto.LoginMember;
-import com.allog.dallog.domain.auth.presentation.AuthenticationPrincipal;
 import com.allog.dallog.domain.subscription.application.SubscriptionService;
+import com.allog.dallog.domain.auth.dto.LoginMember;
 import com.allog.dallog.domain.subscription.dto.request.SubscriptionUpdateRequest;
 import com.allog.dallog.domain.subscription.dto.response.SubscriptionResponse;
 import com.allog.dallog.domain.subscription.dto.response.SubscriptionsResponse;
+import com.allog.dallog.presentation.auth.AuthenticationPrincipal;
 import java.net.URI;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
