+++ backend/src/main/java/com/allog/dallog/presentation/auth/AuthController.java
@@ -1,4 +1,4 @@
-package com.allog.dallog.domain.auth.presentation;
+package com.allog.dallog.presentation.auth;
 
 import com.allog.dallog.domain.auth.application.AuthService;
 import com.allog.dallog.domain.auth.dto.OAuthUriResponse;
