+++ backend/src/test/java/com/allog/dallog/presentation/SubscriptionControllerTest.java
@@ -1,4 +1,4 @@
-package com.allog.dallog.domain.subscription.presentation;
+package com.allog.dallog.presentation;
 
 import static com.allog.dallog.common.fixtures.CategoryFixtures.BE_일정_응답;
 import static com.allog.dallog.common.fixtures.CategoryFixtures.FE_일정_응답;
@@ -29,11 +29,11 @@ import static org.springframework.restdocs.request.RequestDocumentation.pathPara
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.allog.dallog.common.config.TestConfig;
+import com.allog.dallog.domain.subscription.application.SubscriptionService;
 import com.allog.dallog.domain.auth.application.AuthService;
+import com.allog.dallog.common.config.TestConfig;
 import com.allog.dallog.domain.auth.exception.NoPermissionException;
 import com.allog.dallog.domain.category.dto.response.CategoryResponse;
-import com.allog.dallog.domain.subscription.application.SubscriptionService;
 import com.allog.dallog.domain.subscription.domain.Color;
 import com.allog.dallog.domain.subscription.dto.request.SubscriptionUpdateRequest;
 import com.allog.dallog.domain.subscription.dto.response.SubscriptionResponse;
